Rename Nbt to NbtFormat
