+++ src/binstarjs03.AerialOBJ.WpfApp/Services/IOService/SavegameLoader/SavegameLoader.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 
-using binstarjs03.AerialOBJ.Core.Nbt;
+using binstarjs03.AerialOBJ.Core.NbtFormat;
 using binstarjs03.AerialOBJ.WpfApp.Models;
 
 namespace binstarjs03.AerialOBJ.WpfApp.Services.IOService.SavegameLoader;
